requirements.txt
absl-py==0.7.1
astor==0.8.0
Django==2.2.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numpy==1.16.4
Pillow==6.0.0
scipy==1.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
wrapt==1.11.2
